updated controller types and Firesore db impl
Removed the need to accpet a specific model type as a param to the
controller functions. Each controller can now define its own ModelType
for model data as a param (if desired). This line of reasoning could be
extended to the return types for the controller functions as well. For
example, it may be desireable to declare something like a
UsersJSONController, which would accept JSON input and return JSON
output (rather than returning defined models).